Fixed lr and weight_decay being ints

+++ pipelines/scripts/word_alignment/autoencoder.py
@@ -209,8 +209,8 @@ if __name__ == "__main__":
         )
     parser.add_argument("--num-epochs", type=int, help="Number of epochs to train for", default=1)
     parser.add_argument("--batch-size", type=int, help="Batch size for training", default=128)
-    parser.add_argument("--lr", type=int, help="Learning rate for training", default=0.001)
-    parser.add_argument("--weight-decay", type=int, help="Weight decay for training", default=1e-6)
+    parser.add_argument("--lr", type=float, help="Learning rate for training", default=0.001)
+    parser.add_argument("--weight-decay", type=float, help="Weight decay for training", default=1e-6)
     parser.add_argument("--outpath", type=Path, help="where to store results")
     parser.add_argument("--refresh-cache", action='store_true', help="Refresh the cache of match scores")
